GATEWAY-3180: bugfix for refactor of passthru pd def resp.
